remove import of pillar/cmd/tpmmgr
